refactor: 서비스 종류에 맞춰 api 분리
